Move ParseOne to reader.cc
(it is a stand in for the eventual TableReader impl)

Also: do not rely on pointer identity of tags
(`std::shared_ptr<const KeyValueMetadata>`)
